cdk/requirements.txt
aws-cdk.core==1.167.0
aws-cdk.aws-apigatewayv2==1.167.0
aws-cdk.aws-apigatewayv2-integrations==1.167.0
aws-cdk.aws-dynamodb==1.167.0
aws-cdk.aws-s3==1.167.0
aws-cdk.aws-cloudfront==1.167.0
aws-cdk.aws-cloudfront-origins==1.167.0
aws-cdk.aws-lambda==1.167.0
